style: abstract -> PieceType으로 변경함에 따라 테스트 코드 구조 변경
