+++ src/test/java/chess/domain/piece/strategy/BlackPawnMovementStrategyTest.java
@@ -1,4 +1,4 @@
-package chess.domain.piece.type;
+package chess.domain.piece.strategy;
 
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class PawnTest {
+class BlackPawnMovementStrategyTest {
 
     @DisplayName("블랙 폰은 앞으로 한 칸 움직일 수 있다.")
     @Test
@@ -17,21 +17,6 @@ class PawnTest {
         Piece pawn = new Piece(PieceType.BLACK_PAWN);
         Position source = Position.of("a6");
         Position target = Position.of("a5");
-
-        // when
-        boolean result = pawn.isInMovableRange(source, target);
-
-        // then
-        assertThat(result).isTrue();
-    }
-
-    @DisplayName("화이트폰은 앞으로 한 칸 움직일 수 있다.")
-    @Test
-    void canWhitePawnMoveOneStep() {
-        // given
-        Piece pawn = new Piece(PieceType.WHITE_PAWN);
-        Position source = Position.of("a2");
-        Position target = Position.of("a3");
 
         // when
         boolean result = pawn.isInMovableRange(source, target);
@@ -55,21 +40,6 @@ class PawnTest {
         assertThat(result).isTrue();
     }
 
-    @DisplayName("화이트폰은 처음(Rank2) 움직일 때, 앞으로 한 칸  또는 두 칸 움직일 수 있다.")
-    @Test
-    void canWhitePawnMoveTwoStep() {
-        // given
-        Piece pawn = new Piece(PieceType.WHITE_PAWN);
-        Position source = Position.of("a2");
-        Position target = Position.of("a4");
-
-        // when
-        boolean result = pawn.isInMovableRange(source, target);
-
-        // then
-        assertThat(result).isTrue();
-    }
-
     @DisplayName("블랙폰은 처음(Rank7) 움직이는 것이 아닐 때, 앞으로 두 칸 움직일 수 없다.")
     @Test
     void cannotBlackPawnMoveTwoStep() {
@@ -85,36 +55,6 @@ class PawnTest {
         assertThat(result).isFalse();
     }
 
-    @DisplayName("화이트폰은 처음(Rank2) 움직이는 것이 아닐 때, 앞으로 두 칸 움직일 수 없다.")
-    @Test
-    void cannotWhitePawnMoveTwoStep() {
-        // given
-        Piece pawn = new Piece(PieceType.WHITE_PAWN);
-        Position source = Position.of("a3");
-        Position target = Position.of("a5");
-
-        // when
-        boolean result = pawn.isInMovableRange(source, target);
-
-        // then
-        assertThat(result).isFalse();
-    }
-
-    @DisplayName("화이트폰은 앞 측 대각으로 한 칸 움직일 수 있다.")
-    @Test
-    void canWhitePawnMoveDiagonalOneStep() {
-        // given
-        Piece pawn = new Piece(PieceType.WHITE_PAWN);
-        Position source = Position.of("a3");
-        Position target = Position.of("b4");
-
-        // when
-        boolean result = pawn.isInMovableRange(source, target);
-
-        // then
-        assertThat(result).isTrue();
-    }
-
     @DisplayName("블랙폰은 앞 측 대각으로 한 칸 움직일 수 있다.")
     @Test
     void canBlackPawnMoveDiagonalOneStep() {
@@ -128,21 +68,6 @@ class PawnTest {
 
         // then
         assertThat(result).isTrue();
-    }
-
-    @DisplayName("화이트폰은 앞 측 대각으로는 두 칸 움직일 수 없다.")
-    @Test
-    void cannotWhitePawnMoveDiagonalTwoStep() {
-        // given
-        Piece pawn = new Piece(PieceType.WHITE_PAWN);
-        Position source = Position.of("a2");
-        Position target = Position.of("c4");
-
-        // when
-        boolean result = pawn.isInMovableRange(source, target);
-
-        // then
-        assertThat(result).isFalse();
     }
 
     @DisplayName("블랙폰은 앞 측 대각으로는 두 칸 움직일 수 없다.")
